requirements.txt
# Core dependencies
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# AWS integration
boto3==1.34.0
botocore==1.34.0

# Configuration and environment
python-dotenv==1.0.0

# Utility libraries
fake-useragent==1.4.0
tenacity==8.2.3
schedule==1.2.0

# Data handling
python-dateutil==2.8.2

# Development and testing
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.5.0

# Production monitoring
structlog==23.2.0